Добавил документацию, комментарии и сделал рефакторинг
