+++ algorithm/decision-tree.html
@@ -175,7 +175,7 @@
         </table>
 
         <div class="animation disco beat">
-            <button id="load-tests">Классификация</button>
+            <button id="run-classification">Классификация</button>
         </div>
     </div>
